requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
